+++ CHANGELOG.md
@@ -7,9 +7,11 @@ The format of the file is based on a template from [Keep a Changelog](http://kee
 ### Added
 - RoiFileToCSVConverter: operator to write ImageJ ROI data into CSV file
 - DrawRegion2DSet: added some setters to specify dimensions and offsets of target image programmatically
+- MTBQuadraticCurve2D: new methods to extract closest point on ellipse and to calculate tangent orientation at given position
 - SkeletonAnalysisHelper: new class with methods useful when analyzing region skeletons
 
 ### Changed
+- LabelImageEditor: file filter argument is now optional, by default (and with an empty string) all files are processed
 
 ### Deprecated
 
@@ -17,6 +19,7 @@ The format of the file is based on a template from [Keep a Changelog](http://kee
 - Polygon2D_Cgal wrapper class for CGAL due to outdated native bindings
 
 ### Fixed
+- MTBImage, ImageDimensionReducer: preserve calibration information of original images
 - OrientedFilter2D: explicitly shutdown ExecutorService of libimg2 in FFT mode, otherwise operator does not terminate when called from commandline
 
 ## [1.8.10] - 2018-05-18